Make tensorflow imports lazy
