+++ energy_fault_detector/core/__init__.py
@@ -1,8 +1,18 @@
 """This module contains class templates for most of the anomaly detection classes, such as
-autoencoders, anomaly scores, threshold selectors and data classes."""
+anomaly scores, threshold selectors and data classes.
+
+Core templates (no heavy backends (i.e. tensorflow) imported at module import).
+"""
 
 from .anomaly_score import AnomalyScore
-from .autoencoder import Autoencoder
 from .data_transformer import DataTransformer
 from .threshold_selector import ThresholdSelector
 from .fault_detection_result import FaultDetectionResult, ModelMetadata
+
+__all__ = [
+    "AnomalyScore",
+    "DataTransformer",
+    "ThresholdSelector",
+    "FaultDetectionResult",
+    "ModelMetadata",
+]
