+++ energy_fault_detector/root_cause_analysis/__init__.py
@@ -1,3 +1,1 @@
 """Root cause analysis classes"""
-
-from energy_fault_detector.root_cause_analysis.arcana import Arcana
